+++ RELEASE_NOTES.md
@@ -9,6 +9,7 @@ This page describes the noteworthy improvements provided by each release of Ecli
 - inject source folders into maven model
 - improve CI friendly versions
 - Report download speed to the caller
+- redirectTestOutputToFile for OsgiSurefireBooter
 
 ## 3.0.4